Several bug fixes in role trigger custom commands.  (#2005)
* fix bug with incorrect warning on missing trigger channel

* fix bug with incorrect warning on missing trigger channel

* Fixes to Improper use of func to fetch message cc trigger to fetch all type of cc, fix issues with interval ccs triggering when role is being assigned
