+++ customcommands/handle_component.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"sort"
 	"strings"
+	"time"
 
 	"emperror.dev/errors"
 	"github.com/botlabs-gg/yagpdb/v2/bot"
@@ -15,6 +16,7 @@ import (
 	"github.com/botlabs-gg/yagpdb/v2/lib/discordgo"
 	"github.com/botlabs-gg/yagpdb/v2/lib/dstate"
 	"github.com/sirupsen/logrus"
+	"github.com/volatiletech/sqlboiler/v4/queries/qm"
 )
 
 func handleInteractionCreate(evt *eventsystem.EventData) {
@@ -144,8 +146,25 @@ func deferResponseToCCs(interaction *templates.CustomCommandInteraction, ccs []*
 	}
 }
 
+var cachedCommandsComponentTrigger = common.CacheSet.RegisterSlot("custom_commands_component_trigger", nil, int64(0))
+
+func BotCachedGetCommandsWithComponentTrigger(guildID int64, ctx context.Context) ([]*models.CustomCommand, error) {
+	v, err := cachedCommandsComponentTrigger.GetCustomFetch(guildID, func(key interface{}) (interface{}, error) {
+		var cmds []*models.CustomCommand
+		var err error
+		common.LogLongCallTime(time.Second, true, "Took longer than a second to fetch custom commands from db", logrus.Fields{"guild": guildID}, func() {
+			cmds, err = models.CustomCommands(qm.Where("guild_id = ? AND trigger_type IN (7,8)", guildID), qm.OrderBy("local_id desc"), qm.Load("Group")).AllG(ctx)
+		})
+		return cmds, err
+	})
+	if err != nil {
+		return nil, err
+	}
+	return v.([]*models.CustomCommand), nil
+}
+
 func findComponentOrModalTriggerCustomCommands(ctx context.Context, cs *dstate.ChannelState, member *discordgo.Member, cID string, modal bool) (matches []*TriggeredCC, err error) {
-	cmds, err := BotCachedGetCommandsWithMessageTriggers(cs.GuildID, ctx)
+	cmds, err := BotCachedGetCommandsWithComponentTrigger(cs.GuildID, ctx)
 	if err != nil {
 		return nil, errors.WrapIf(err, "BotCachedGetCommandsWithComponentTriggers")
 	}
@@ -264,6 +283,90 @@ func ExecuteCustomCommandFromComponent(cc *models.CustomCommand, gs *dstate.Guil
 func CheckMatchComponent(cmd *models.CustomCommand, cID string) (match bool, stripped string, args []string) {
 
 	if cmd.TriggerType != int(CommandTriggerComponent) {
+		return false, "", nil
+	}
+
+	cmdMatch := "(?m)"
+	if !cmd.TextTriggerCaseSensitive {
+		cmdMatch += "(?i)"
+	}
+	cmdMatch += cmd.TextTrigger
+
+	match, stripped, args = matchRegexSplitArgs(cmdMatch, cID)
+	return
+}
+
+func ExecuteCustomCommandFromModal(cc *models.CustomCommand, gs *dstate.GuildSet, cs *dstate.ChannelState, cmdArgs []string, stripped string, interaction *templates.CustomCommandInteraction) error {
+	ms := dstate.MemberStateFromMember(interaction.Member)
+	tmplCtx := templates.NewContext(gs, cs, ms)
+	tmplCtx.CurrentFrame.Interaction = interaction
+
+	tmplCtx.Data["Interaction"] = interaction
+	tmplCtx.Data["InteractionData"] = interaction.ModalSubmitData()
+	modalCustomID := strings.TrimPrefix(interaction.ModalSubmitData().CustomID, templates.TemplateCustomIDPrefix)
+	tmplCtx.Data["CustomID"] = modalCustomID
+	tmplCtx.Data["Cmd"] = cmdArgs[0]
+	if len(cmdArgs) > 1 {
+		tmplCtx.Data["CmdArgs"] = cmdArgs[1:]
+	} else {
+		tmplCtx.Data["CmdArgs"] = []string{}
+	}
+	tmplCtx.Data["StrippedID"] = stripped
+	tmplCtx.Data["StrippedMsg"] = stripped
+	tmplCtx.Data["IsModal"] = true
+	cmdValues := []any{}
+
+	modalValues := templates.SDict{}
+	for i := 0; i < len(interaction.ModalSubmitData().Components); i++ {
+		switch comp := interaction.ModalSubmitData().Components[i].(type) {
+		case *discordgo.ActionsRow:
+			for j := 0; j < len(comp.Components); j++ {
+				field, ok := comp.Components[j].(*discordgo.TextInput)
+				if ok {
+					cmdValues = append(cmdValues, field.Value)
+				}
+				cID, _ := strings.CutPrefix(field.CustomID, templates.TemplateCustomIDPrefix)
+				modalValues.Set(cID, templates.SDict{
+					"type":      field.Type(),
+					"value":     field.Value,
+					"custom_id": cID,
+				})
+			}
+		case *discordgo.Label:
+			if t, ok := comp.Component.(*discordgo.TextInput); ok {
+				cID, _ := strings.CutPrefix(t.CustomID, templates.TemplateCustomIDPrefix)
+				cmdValues = append(cmdValues, t.Value)
+				modalValues.Set(cID, templates.SDict{
+					"type":      t.Type(),
+					"value":     t.Value,
+					"custom_id": cID,
+				})
+			} else if sm, ok := comp.Component.(*discordgo.SelectMenu); ok {
+				cID, _ := strings.CutPrefix(sm.CustomID, templates.TemplateCustomIDPrefix)
+				cmdValues = append(cmdValues, sm.Values)
+				modalValues.Set(cID, templates.SDict{
+					"type":      sm.Type(),
+					"value":     sm.Values,
+					"custom_id": cID,
+				})
+			}
+		}
+	}
+	tmplCtx.Data["Values"] = cmdValues
+	tmplCtx.Data["ModalValues"] = modalValues
+	msg := interaction.Message
+	msg.Member = ms.DgoMember()
+	msg.Author = msg.Member.User
+	tmplCtx.Msg = msg
+
+	tmplCtx.Data["Message"] = msg
+
+	return ExecuteCustomCommand(cc, tmplCtx)
+}
+
+func CheckMatchModal(cmd *models.CustomCommand, cID string) (match bool, stripped string, args []string) {
+
+	if cmd.TriggerType != int(CommandTriggerModal) {
 		return false, "", nil
 	}
